refactor: entity logs
